AP_InertialSensor: fix sign-compare in QMI8658A/HXY42688 regs check (SITL -Werror)
Vendor drivers used 'for (int i = 0; i < sizeof(COMM_REGS)/sizeof(...))' — signed
counter vs size_t triggers -Wsign-compare, fatal under SITL's -Werror (the ARM
minimized build didn't error). Use size_t i. Behavior identical on ARM; the build
branch now SITL-builds so it can be autotested directly.

+++ libraries/AP_InertialSensor/AP_InertialSensor_HXY42688.cpp
@@ -195,7 +195,7 @@ bool AP_InertialSensor_HXY42688::_regs_check_out()
     write_register(0x7F, 0x00);//goto 0x00
     hal.scheduler->delay(1);
 
-    for (int i = 0; i < sizeof(AP_InertialSensor_HXY42688::COMM_REGS)/sizeof(AP_InertialSensor_HXY42688::COMM_REGS[0]); ++i) {
+    for (size_t i = 0; i < sizeof(AP_InertialSensor_HXY42688::COMM_REGS)/sizeof(AP_InertialSensor_HXY42688::COMM_REGS[0]); ++i) {
         uint8_t reg = AP_InertialSensor_HXY42688::COMM_REGS[i][0];
         uint8_t value = read_register(reg);
         if (reg == 0x40 || reg == 0x42) {
